test: collect sample tasks in one class
